GEODE-5742: Restrict default membership range to not conflict with default ports
Currently the default range for membership ports is 1024-65535.  This results
in occassional failures where a random membership port selection will conflict
with the default server or locator port when using all default values.

This change modifies the default range to 41000-61000.  In addition to not
conflicting with any of the default ports (highest is 40404), this is also a
proper subset of the default Linux ephemeral port range (32768-61000).